Move test utility library to the util directory
